feat: ajout de la gestion des numéros de téléphone et mise à jour des rôles d'utilisateur

+++ client/src/admin.ts
@@ -1,4 +1,9 @@
 import './style.css';
+const section_id = document.querySelector<HTMLDivElement>("#section_id");
+const name = localStorage.getItem("name");
+const firstname = localStorage.getItem("firstname");
+const id = localStorage.getItem("id");
+if (section_id) section_id.textContent = `${name} ${firstname} ID : ${id}`;
 const button_deconnexion = document.querySelector<HTMLButtonElement>("#deconnexion");
 const out = document.querySelector<HTMLPreElement>("#out");
 button_deconnexion?.addEventListener("click", async (e) => {
@@ -71,30 +76,70 @@ button_deconnexion?.addEventListener("click", async (e) => {
 
         })
 
-        const role = document.createElement("div");
+        const role = document.createElement("button");
         role.className = "card-role";
         role.textContent = `Role : ${u.role}`;
-        role.classList.add(u.role === "admin" ? "role-admin" : "role-user");
+        role.classList.add(`role-${u.role}`);
+
+        role.addEventListener("click", async (e) => {
+            e.preventDefault();
+            const payload = {
+                role: u.role
+            };
+
+            try {
+                const res = await fetch(`/api/admin/set_approved/${u.id}`, {
+                    method: "PATCH",
+                    headers: { "Content-Type": "application/json" },
+                    credentials: 'include',
+                    body: JSON.stringify(payload),
+                })
+                const json = await res.json();
+                if (!res.ok) {
+                    if (out) out.textContent = `Message Erreur : ${json.message}`;
+                    return
+                } else {
+
+                    if (out) out.textContent = `${json.message}`;
+                    role.textContent = `Role : ${json.data.role}`;
+                    role.classList.remove("role-admin", "role-user", "role-visitor");
+                    role.classList.add(`role-${json.data.role}`);
+                    u.role = json.data.role;
+
+
+                }
+            } catch (error) {
+                if (out) out.textContent = `Error : ${error}`;
+                return
+            }
+
+
+        })
 
         const button_delete = document.createElement("button");
         button_delete.className = "card-button_delete";
         button_delete.textContent = `Suppression`;
-
         button_delete.addEventListener("click", async (e) => {
             e.preventDefault();
-            const res = await fetch(`/api/admin/delete/${u.id}`, {
-                method: "DELETE",
-                headers: { "Content-Type": "application/json" },
-                credentials: 'include'
-            })
-            if (!res.ok) {
-                const json = await res.json();
-                if (out) out.textContent = json.message;
-            } else {
-                const json = await res.json();
-                if (out) out.textContent = json.message;
-                alert(json.message);
-                card.remove();
+            if (confirm(`Voulez vous réellement supprimer l'utilisateur : ${u.name} ${u.firstname}`)) {
+                const res = await fetch(`/api/admin/delete/${u.id}`, {
+                    method: "DELETE",
+                    headers: { "Content-Type": "application/json" },
+                    credentials: 'include'
+                })
+                if (!res.ok) {
+                    const json = await res.json();
+                    if (out) out.textContent = json.message;
+                } else {
+
+                    const json = await res.json();
+
+                    if (out) out.textContent = json.message;
+                    alert(json.message);
+                    card.remove();
+
+
+                }
             }
 
         })
